move Int subclass to jdaviz.core.custom_traitlets
* and add test coverage to catch any breaks from upstream changes
